usersPolicy

+++ app/Http/Controllers/MessageContentController.php
@@ -3,7 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\MessageContent;
+use App\Models\Chat;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use App\Models\UserIntermediary;
+use App\Models\MessageType;
 use Validator;
 
 
@@ -38,14 +42,34 @@ class MessageContentController extends Controller
      */
     public function store(Request $request)
     {
-         $rules = [
-            'type_id'=>'required|integer',
+
+        $user = Auth::user();
+
+        $serverId = Chat::select('server_id')->where('chat_id',$request->chat_id)->first();
+
+        if(empty($serverId)){
+            return response()->json(['error'=>true,'message'=>'Chat not found']);
+        }
+        $check = UserIntermediary::select('id')->where('user_id',$user['user_id'])->where('server_id', $serverId['serverId']);
+
+        if(empty($check)){
+            return response()->json(['error'=> true, 'message' => 'User not found in server'],404);
+        }
+
+        $typeMessage = MessageType::select('id')->where('type', $request->type_message)->first();
+
+        if(empty($check)){
+            return response()->json(['error'=> true, 'message' => 'Type not found'],404);
+        }
+
+
+        $rules = [
+            'type_message'=>'required',
             'content'=>'required|min:1',
-            'status'=>'required|min:1|max:1|boolean',
+            
             'date'=>'required|date',
             'time'=>'required',
             'chat_id'=>'required|integer',
-            'user_id'=>'required|integer',
         ];
 
         $validator = Validator::make($request->all(),$rules);
@@ -53,7 +77,8 @@ class MessageContentController extends Controller
             return response()->json($validator->errors(),400);
         }
 
-        $messageContent = MessageContent::create($request->all());
+        $messageContent = MessageContent::create(['type_id' => $typeMessage['id'],'content' => $request->content, 'status' => true,'date' => $request->date,'time' => $request->time,'chat_id' => $request->chat_id,'user_id' => $user["user_id"]]);
+
         return response()->json($messageContent,201);
         
     }
@@ -83,6 +108,19 @@ class MessageContentController extends Controller
      */
     public function update(Request $request, $id)
     {
+
+        $user = Auth::user();
+
+        $message = MessageContent::find($id);
+
+        if(is_null($message)){
+            return response()->json(['error'=>true,'message'=>'Message not Found'],404);
+        }
+
+        if($user['user_id'] !== $messageContent['user_id']){
+            return response()->json(['error'=>true,'message'=>'Forbidden'],403);
+        }
+
         $rules = [
             'type_id'=>'required|integer',
             'content'=>'required|min:1',
@@ -98,10 +136,6 @@ class MessageContentController extends Controller
             return response()->json($validator->errors(),400);
         }
 
-        $messageContent = MessageContent::find($id);
-        if(is_null($messageContent)){
-            return response()->json(['error'=>true,'message'=>'Not Found'],404);
-        }
         
         $messageContent->update($request->all());
         return response()->json($messageContent,200);
@@ -115,12 +149,17 @@ class MessageContentController extends Controller
      */
     public function destroy($id)
     {
+        $user = Auth::user();
         $messageContent = MessageContent::find($id);
         if(is_null($messageContent)){
             return response()->json(['error'=>true,'message'=>'Not Found'],404);
         }
+        if($user['user_id'] !== $messageContent['user_id']){
+            return response()->json(['error'=>true,'message'=>'Forbidden'],403);
+        }
+
         $messageContent->delete();
        
-        return response()->json('',204);
+        return response()->json('Deleted',204);
     }
 }
