+++ app/Http/Controllers/UserIntermediariesController.php
@@ -7,6 +7,8 @@ use App\Models\UserIntermediary;
 use App\Models\Server;
 use App\Models\Users;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
+use Validator;
 
 class UserIntermediariesController extends Controller
 {
@@ -38,6 +40,28 @@ class UserIntermediariesController extends Controller
         if($validator->fails()){
             return response()->json($validator->errors(),400);
         }
+        $user = Auth::user();
+
+        $server = Server::find($request->server_id);
+
+        if($user['user_id'] !== $server['admin_id']){
+            return response()->json(['error'=>true,'message'=>'Forbidden'],403);
+        } 
+
+        $checkUser = Users::find($request->user_id);
+
+        if(!empty($check)){
+            return response()->json(['error'=>true,'message'=>'User not found']);
+        } 
+
+        $check = UserIntermediary::select('id')->where('server_id',2)->where('user_id', $request->user_id)->first();
+        //return response()->json($check);
+     //return response()->json(empty($check));
+
+        if(!empty($check)){
+            return response()->json(['message'=>'Resourse exists']);
+        } 
+
 
         $userIntermediary = UserIntermediary::create($request->all());
         return response()->json($userIntermediary,201);
@@ -59,19 +83,19 @@ class UserIntermediariesController extends Controller
         return response()->json($userIntermediary,200);
     }
 
-    public function showUserId($id){
+    public function showUserId(){
         $servers = [];
-        $user = Users::find($id);
+        $user = Auth::user();
         if(is_null($user)){
             return response()->json(['error'=>true,'message'=>'Not Found'],404);}
-        $userIntermediary = DB::table('user_intermediaries')->select('server_id')->where('user_id',"$id")->get();
+        $userIntermediary = DB::table('user_intermediaries')->select('server_id')->where('user_id',$user['user_id'])->get();
 
          foreach ($userIntermediary as $value) {
             $servers[] = Server::find($value->server_id); 
         }
         
         
-        return response()->json(['user_id' =>$id, 'servers'=>$servers]);
+        return response()->json(['user_id' =>$user['user_id'], 'servers'=>$servers]);
     }
 
     public function showServerId($id){
@@ -133,10 +157,31 @@ class UserIntermediariesController extends Controller
         // $server = Server::find($id);
         // if(is_null($server)){
         //     return response()->json(['error'=>true,'message'=>'Not Found'],404);}
-         $userIntermediary = DB::table('user_intermediaries')->select('id')->where('user_id',"$user_id")->where('server_id',"$server_id")->first();
+        //     
+        $user = Auth::user();
+        $server = Server::find($server_id);
+
+        if(empty($server)){
+             return response()->json(['error'=>true,'message'=>'Not found server'],404);
+        }
+
+        if($user['user_id'] !== $server['admin_id']){
+            return response()->json(['error'=>true,'message'=>'Forbidden'],403);
+        }  
+
+        if($user['user_id'] === (int)$user_id){
+            return response()->json(['error'=>true,'message'=>'Not delete admin'],403);
+        }
+
+
+        $userIntermediary = DB::table('user_intermediaries')->select('id')->where('user_id',$user_id)->where('server_id',$server['server_id'])->first();
+
+        if(empty($userIntermediary)){
+            return response()->json(['error'=>true,'message'=>'User not found in server'],404);
+        }
 
         DB::table('user_intermediaries')->where([['user_id', $user_id],['server_id',$server_id]])->delete();
        
-        return response()->json('',204);
+        return response()->json('User delete from server',204);
     }
 }
